+++ saleor/graphql/warehouse/mutations/base.py
@@ -1,6 +1,5 @@
 from django.core.exceptions import ValidationError
 
-from ...account.i18n import I18nMixin
 from ....warehouse import WarehouseClickAndCollectOption
 from ....warehouse.error_codes import WarehouseErrorCode
 from ...core import ResolveInfo
@@ -10,7 +9,7 @@ from ...core.validators import (
 )
 
 
-class WarehouseMixin(I18nMixin):
+class WarehouseMixin:
     @classmethod
     def clean_input(cls, info: ResolveInfo, instance, data, **kwargs):
         cleaned_input = super().clean_input(  # type: ignore[misc] # mixin
@@ -59,24 +58,4 @@ class WarehouseMixin(I18nMixin):
                     )
                 },
             )
-        cleaned_input["address"] = cls.prepare_address(cleaned_input, info)
         return cleaned_input
-
-    @classmethod
-    def prepare_address(cls, cleaned_input, info):
-        address_data = cleaned_input.get("address", {})
-        if not address_data:
-            # TODO zedzior check it
-            return None
-        address_instance = cls.validate_address(address_data, info=info)
-        # address_metadata = address_data.pop("metadata", list())
-        # if address_metadata:
-        #     address_instance.store_value_in_metadata(
-        #         {data.key: data.value for data in address_metadata}
-        #     )
-        return address_instance
-
-    @classmethod
-    def construct_instance(cls, instance, cleaned_data):
-        cls.update_metadata(instance, cleaned_data.pop("metadata", list()))  # type: ignore[attr-defined] # noqa: E501
-        return cls.construct_instance(instance, cleaned_data)  # type: ignore[misc] # noqa: E501
