+++ saleor/graphql/warehouse/mutations/warehouse_create.py
@@ -10,7 +10,7 @@ from ..types import Warehouse, WarehouseCreateInput
 from .base import WarehouseMixin
 
 
-class WarehouseCreate(AddressMetadataMixin, WarehouseMixin, ModelMutation):
+class WarehouseCreate(AddressMetadataMixin, WarehouseMixin, ModelMutation, I18nMixin):
     class Arguments:
         input = WarehouseCreateInput(
             required=True, description="Fields required to create warehouse."
@@ -25,6 +25,28 @@ class WarehouseCreate(AddressMetadataMixin, WarehouseMixin, ModelMutation):
         error_type_field = "warehouse_errors"
 
     @classmethod
+    def prepare_address(cls, cleaned_input, info):
+        address_data = cleaned_input.get("address", {})
+        if not address_data:
+            # TODO zedzior test it
+            return None
+        address_instance = cls.validate_address(address_data, info=info)
+        address_instance.save()
+        return address_instance
+
+    @classmethod
     def post_save_action(cls, info: ResolveInfo, instance, cleaned_input):
         manager = get_plugin_manager_promise(info.context).get()
         cls.call_event(manager.warehouse_created, instance)
+
+    @classmethod
+    def perform_mutation(cls, _root, info: ResolveInfo, /, **data):
+        instance = cls.get_instance(info, **data)
+        data = data.get("input")
+        cleaned_input = cls.clean_input(info, instance, data)
+        cleaned_input["address"] = cls.prepare_address(cleaned_input, info)
+        instance = cls.construct_instance(instance, cleaned_input)
+        cls.clean_instance(info, instance)
+        cls.save(info, instance, cleaned_input)
+        cls.post_save_action(info, instance, cleaned_input)
+        return cls.success_response(instance)
